refactor: Restructure project into MVC+SR pattern
- Changed folder structure to follow MVC + SR (Service/Repository) architecture
- Moved logic from feature-based folders (gameflow, input, output) into controller, service, view, model
- Updated all package names to match the new folder structure
- Fixed all import statements to match the updated packages
- Moved GameConfig from controller to model to keep clean dependency direction
- All tests passed after refactor (./gradlew clean test)
