+++ src/main/kotlin/racingcar/repository/README.md
@@ -1,0 +1,3 @@
+# Repository Layer (Reserved)
+
+This directory is reserved for future use in case persistent storage(ex: DB) is added. For more details, you can check the information about Repository Layer in the Hierarchy section of the project README.md file.
