+++ src/main/kotlin/racingcar/service/raceexecution/RaceMoveCars.kt
@@ -1,4 +1,4 @@
-package racingcar.raceexecution
+package racingcar.service.raceexecution
 
 /**
  *  [ Goal ]
